# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentation de l'API KoriPay",
        "description": "Interface interactive pour tester les endpoints de la plateforme financière KoriPay (Projet L3 IAGE / ISI Dakar).",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/",
            "description": "Serveur Local de Développement"
        }
    ],
    "paths": {
...
        "/recuperation/verifier-identite": {
            "post": {
                "tags": [
                    "Récupération de compte"
                ],
                "summary": "PIN oublié Étape 1 : Audit des habitudes financières",
                "description": "Vérifie l'identité du client via ses habitudes de transactions et sa question secrète. Accessible uniquement depuis un appareil déjà enregistré. En cas de succès, retourne un ticket d'autorisation valide 5 minutes. Après 3 échecs consécutifs, le compte est temporairement suspendu 30 minutes.",
                "operationId": "recuperationVerifierIdentite",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "telephone",
                                    "device_id",
                                    "montant_derniere_tx",
                                    "contact_frequent"
                                ],
                                "properties": {
                                    "telephone": {
                                        "type": "string",
                                        "example": "771111111"
                                    },
                                    "device_id": {
                                        "type": "string",
                                        "example": "f7b129a0-c3bc"
                                    },
                                    "montant_derniere_tx": {
                                        "description": "Montant exact de la toute dernière transaction",
                                        "type": "number",
                                        "example": 15000
                                    },
                                    "contact_frequent": {
                                        "description": "Numéro ou nom du contact avec qui vous avez le plus de transactions (envoi ET réception)",
                                        "type": "string",
                                        "example": "771234567"
                                    },
                                    "reponse_secrete": {
                                        "description": "Requis uniquement si une question secrète a été configurée",
                                        "type": "string",
                                        "example": "Dakar",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Identité confirmée, ticket d'autorisation généré"
                    },
                    "400": {
                        "description": "Appareil non reconnu, ou informations complémentaires requises (contact fréquent ou question secrète manquante)"
                    },
                    "401": {
                        "description": "Réponse incorrecte (montant, contact fréquent, ou réponse secrète) - indique le nombre de tentatives restantes avant suspension"
                    },
                    "403": {
                        "description": "Compte temporairement suspendu suite à 3 échecs consécutifs"
                    },
                    "404": {
                        "description": "Compte introuvable"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
        "/recuperation/reinitialiser-pin": {
            "post": {
                "tags": [
                    "Récupération de compte"
                ],
                "summary": "PIN oublié Étape 2 : Réinitialiser le code PIN",
                "description": "Applique le nouveau code PIN après validation du ticket d'autorisation obtenu à l'étape 1. Le ticket est à usage unique et expire dans 5 minutes. Après 3 échecs consécutifs, le compte est temporairement suspendu 30 minutes.",
                "operationId": "recuperationReinitialiserPin",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "telephone",
                                    "ticket",
                                    "nouveau_pin",
                                    "nouveau_pin_confirmation"
                                ],
                                "properties": {
                                    "telephone": {
                                        "type": "string",
                                        "example": "771111111"
                                    },
                                    "ticket": {
                                        "description": "Ticket reçu à l'étape 1",
                                        "type": "string",
                                        "example": "482910"
                                    },
                                    "nouveau_pin": {
                                        "description": "Nouveau code PIN à 4 chiffres",
                                        "type": "string",
                                        "example": "4321"
                                    },
                                    "nouveau_pin_confirmation": {
                                        "description": "Confirmation du nouveau code PIN",
                                        "type": "string",
                                        "example": "4321"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Code PIN réinitialisé avec succès"
                    },
                    "401": {
                        "description": "Ticket invalide ou expiré - indique le nombre de tentatives restantes avant suspension"
                    },
                    "403": {
                        "description": "Compte temporairement suspendu suite à 3 échecs consécutifs"
                    },
                    "404": {
                        "description": "Compte introuvable"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Entrez votre jeton de sécurité (Token Sanctum) reçu lors du Login pour tester les routes protégées.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}